Create custom pagiinition and add to datagrid
